Should have fixed many more signed/unsigned warnings
Also fixed an issue with NDFA accepting actions: VC++ resolves integers as booleans
and doesn't call any implicit constructors.

--HG--
branch : Trunk
extra : transplant_source : %99%A2%B1%FF_%B9%22%C00r%B3%F8%D2%88%20n%07%20J%01

+++ TameParse/Lr/conflict.cpp
@@ -150,7 +150,7 @@ static void find_conflicts(const lalr_builder& builder, int stateId, conflict_li
         }
 
         // Iterate through the items in the state to look for the reduce actions
-        for (int itemId = 0; itemId < thisState.count_items(); itemId++) {
+        for (size_t itemId = 0; itemId < thisState.count_items(); itemId++) {
             // Get the item that this refers to
             const lalr_state::container& thisItem = thisState[itemId];
